Text.Email.Validate

Documentation

addrSpec :: Parsec String () Int

isValid :: String -> Bool

Validates whether a particular string is an email address according to RFC5322.

validate :: String -> Either ParseError EmailAddress

If you want to find out why a particular string is not an email address, use this!

data EmailAddress

Constructor does no checking for invalid emails, so use at own risk.

Constructors

EmailAddress

Fields

localPart :: String

domainPart :: String