Text.Email.Validate

Documentation

isValid :: ByteString -> Bool

Validates whether a particular string is an email address according to RFC5322.

validate :: ByteString -> Either String EmailAddress

If you want to find out *why* a particular string is not an email address, use this.

emailAddress :: ByteString -> Maybe EmailAddress

Smart constructor for an email address

canonicalizeEmail :: ByteString -> Maybe ByteString

Checks that an email is valid and returns a version of it where comments and whitespace have been removed.

data EmailAddress

Represents an email address.

unsafeEmailAddress :: ByteString -> ByteString -> EmailAddress

Creates an email address without validating it. You should only use this when reading data from somewhere it has already been validated (e.g. a database).

localPart :: EmailAddress -> ByteString

Extracts the local part of an email address.

domainPart :: EmailAddress -> ByteString

Extracts the domain part of an email address.

toByteString :: EmailAddress -> ByteString

Converts an email address back to a ByteString